requirements.txt
streamlit==1.32.0
pdfplumber==0.10.4
PyPDF2==3.0.1
pytesseract==0.3.10
pandas==2.2.1
openpyxl==3.1.2
google-generativeai==0.4.1
sentence-transformers==2.5.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.121.0
python-dotenv==1.0.1
plotly==5.19.0
dnspython==2.6.1
comet-ml==3.38.1
reportlab==4.1.0
pytest==8.0.2
python-multipart==0.0.9
watchdog==4.0.0
pdf2image==1.17.0
